+++ server/feature/game/games.go
@@ -35,7 +35,10 @@ func (s *Service) saveGame(c *entity.Game, onlyUpdate bool) error {
 		return errors.New("game missing id or title")
 	}
 	if c.CoverID != "" {
-		p, err := image.DownloadAndInsertImage(s.db, "https://images.igdb.com/igdb/image/upload/t_cover_big/"+c.CoverID+".png", "games")
+		p, err := image.DownloadAndInsertFromUrl(
+			s.db,
+			"https://images.igdb.com/igdb/image/upload/t_cover_big/"+c.CoverID+".png",
+			"games")
 		if err != nil {
 			slog.Error("saveGame: Failed to cache game cover.", "error", err)
 		} else {