+++ server/util/size.go
@@ -1,0 +1,7 @@
+package util
+
+const (
+	Byte     int64 = 1
+	Kibibyte int64 = 1024
+	Mebibyte int64 = 1024 * Kibibyte
+)